Zero-Sum Games

Game Problem

OperationsResearchModels.Game.GameProblem — Type

GameProblem

Description

Defines the problem of a zero-sum game.

Fields

decisionMatrix::Matrix{<:Real}: The payoff matrix of the game designed for the row player.

Game solver for the row player

OperationsResearchModels.solve — Method

solve(p::GameProblem; verbose::Bool = false)::Vector{GameResult}

Solves a zero-sum game using the simplex method.

Arguments

p::GameProblem: The problem instance containing the decision matrix.

verbose: If true, prints the model information.

Returns

An array of GameResult objects containing the probabilities and value of the game.

GameResult

OperationsResearchModels.Game.GameResult — Type

GameResult

Description

A structure to hold the result of a game.

Fields

probabilities: Probabilities of the strategies

value: Value of the game

end