pmedian

OperationsResearchModels.PMedian.pmedian — Function

pmedian(data, ncenters)

Arguments

data::Matrix: Coordinates of locations

ncenters::Int: Number of centers

Description

The function calculates Euclidean distances between all possible rows of the matrix data. ncenters locations are then selected that minimizes the total distances to the nearest rows.

Output

PMedianResult: PMedianResult object.

Example

julia> data1 = rand(10, 2);

julia> data2 = rand(10, 2) .+ 50;

julia> data3 = rand(10, 2) .+ 100;

julia> data = vcat(data1, data2, data3);

julia> result = pmedian(data, 3);

julia> result.centers
3-element Vector{Int64}:
  1
 16
 21

 julia> result.objective
 11.531012240599605

pmedian with distances

OperationsResearchModels.PMedian.pmedian_with_distances — Function

pmedian_with_distances(distancematrix, ncenters)

Arguments

distancematrix::Matrix: n x n matrix of distances

ncenters::Int: Number of centers

Description

ncenters locations are selected that minimizes the total distances to the nearest rows.

Output

PMedianResult: PMedianResult object.